+++ src/main/java/com/kaba4cow/futuresscreenerbot/properties/screener/LongLiquidationScreenerSettingsProperties.java
@@ -1,7 +1,5 @@
 package com.kaba4cow.futuresscreenerbot.properties.screener;
 
-import java.math.BigDecimal;
-
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -14,8 +12,8 @@ import lombok.Setter;
 @Component
 public class LongLiquidationScreenerSettingsProperties {
 
-	private BigDecimal minLongLiquidationThreshold;
+	private Double minLongLiquidationThreshold;
 
-	private BigDecimal maxLongLiquidationThreshold;
+	private Double maxLongLiquidationThreshold;
 
 }
