+++ src/main/java/com/kaba4cow/futuresscreenerbot/service/domain/settingsextractor/SubscriberSettingsExtractor.java
@@ -1,7 +1,5 @@
 package com.kaba4cow.futuresscreenerbot.service.domain.settingsextractor;
 
-import java.math.BigDecimal;
-
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.kaba4cow.futuresscreenerbot.entity.EventType;
@@ -14,7 +12,7 @@ public interface SubscriberSettingsExtractor {
 		registry.register(this);
 	}
 
-	BigDecimal getThreshold(SubscriberSettings settings);
+	Double getThreshold(SubscriberSettings settings);
 
 	EventType getEventType();
