+++ src/main/java/com/kaba4cow/futuresscreenerbot/service/domain/settingsextractor/impl/DumpThresholdExtractor.java
@@ -1,7 +1,5 @@
 package com.kaba4cow.futuresscreenerbot.service.domain.settingsextractor.impl;
 
-import java.math.BigDecimal;
-
 import org.springframework.stereotype.Component;
 
 import com.kaba4cow.futuresscreenerbot.entity.EventType;
@@ -12,7 +10,7 @@ import com.kaba4cow.futuresscreenerbot.service.domain.settingsextractor.Subscrib
 public class DumpThresholdExtractor implements SubscriberSettingsExtractor {
 
 	@Override
-	public BigDecimal getThreshold(SubscriberSettings settings) {
+	public Double getThreshold(SubscriberSettings settings) {
 		return settings.getDumpThreshold();
 	}
